+++ app/src/main/java/woowacourse/shopping/ui/detail/DetailUiState.kt
@@ -5,6 +5,6 @@ import woowacourse.shopping.ui.model.ProductUiModel
 data class DetailUiState(
     val product: ProductUiModel = ProductUiModel(),
     val quantity: Int = 1,
-    val totalPrice: Int = 0,
+    val totalPrice: Long = 0,
     val recentItem: ProductUiModel? = null,
 )
